+++ src/browser/components/index.ts
@@ -1,0 +1,5 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+export * from './splashscreen';
+export * from './servermanager';
