+++ src/browser/components/servermanager/servermanager.tsx
