+++ src/browser/components/servermanager/style/servermanager.css
