+++ src/browser/components/splashscreen/index.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+import './style/index.css';
+
+export * from './splash';
