+++ src/browser/components/splashscreen/splash.tsx
