+++ src/browser/components/splashscreen/style/index.css
@@ -3,5 +3,4 @@
 | Distributed under the terms of the Modified BSD License.
 |----------------------------------------------------------------------------*/
 
-@import './servermanager.css';
 @import './splash.css';
