+++ src/browser/components/splashscreen/style/splash.css
