--- src/browser/electron-launcher/index.ts
@@ -1,7 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-
-import 'jupyterlab_app/src/browser/electron-launcher/style/index.css';
-
-export * from 'jupyterlab_app/src/browser/electron-launcher/splash';
-export * from 'jupyterlab_app/src/browser/electron-launcher/servermanager';
