+++ src/browser/extensions/electron-extension/index.ts
