+++ src/browser/index.tsx
@@ -1,7 +1,7 @@
 
 import 'font-awesome/css/font-awesome.min.css';
 import '@jupyterlab/theming/style/index.css';
-import 'jupyterlab_app/src/browser/css/main.css';
+import 'jupyterlab_app/src/browser/style/main.css';
 
 import {
     Application
