+++ src/browser/style/fonts/Cousine/Cousine-Italic.ttf
