+++ src/browser/style/main.css
@@ -24,56 +24,56 @@ table { border-collapse: collapse; border-spacing: 0; }
  
 @font-face {
     font-family: 'Roboto';
-    src: url('../fonts/Roboto/Roboto-Regular.ttf') format('truetype');
+    src: url('fonts/Roboto/Roboto-Regular.ttf') format('truetype');
     font-weight: normal;
     font-style: normal;
 }
 
 @font-face {
     font-family: 'Roboto';
-    src: url('../fonts/Roboto/Roboto-Italic.ttf') format('truetype');
+    src: url('fonts/Roboto/Roboto-Italic.ttf') format('truetype');
     font-weight: normal;
     font-style: italic;
 }
 
 @font-face {
     font-family: 'Roboto';
-    src: url('../fonts/Roboto/Roboto-Bold.ttf') format('truetype');
+    src: url('fonts/Roboto/Roboto-Bold.ttf') format('truetype');
     font-weight: bold;
     font-style: normal;
 }
 
 @font-face {
     font-family: 'Roboto';
-    src: url('../fonts/Roboto/Roboto-BoldItalic.ttf') format('truetype');
+    src: url('fonts/Roboto/Roboto-BoldItalic.ttf') format('truetype');
     font-weight: bold;
     font-style: italic;
 }
 
 @font-face {
     font-family: 'Cousine';
-    src: url('../fonts/Cousine/Cousine-Regular.ttf') format('truetype');
+    src: url('fonts/Cousine/Cousine-Regular.ttf') format('truetype');
     font-weight: normal;
     font-style: normal;
 }
 
 @font-face {
     font-family: 'Cousine';
-    src: url('../fonts/Cousine/Cousine-Italic.ttf') format('truetype');
+    src: url('fonts/Cousine/Cousine-Italic.ttf') format('truetype');
     font-weight: normal;
     font-style: italic;
 }
 
 @font-face {
     font-family: 'Cousine';
-    src: url('../fonts/Cousine/Cousine-Bold.ttf') format('truetype');
+    src: url('fonts/Cousine/Cousine-Bold.ttf') format('truetype');
     font-weight: bold;
     font-style: normal;
 }
 
 @font-face {
     font-family: 'Cousine';
-    src: url('../fonts/Cousine/Cousine-BoldItalic.ttf') format('truetype');
+    src: url('fonts/Cousine/Cousine-BoldItalic.ttf') format('truetype');
     font-weight: bold;
     font-style: italic;
 }
